+++ packages/emails/index.ts
@@ -1,2 +1,8 @@
 export * from "./email-manager";
+export * from "./auth-email-service";
+export * from "./organization-email-service";
+export * from "./billing-email-service";
+export * from "./integration-email-service";
+export * from "./workflow-email-service";
+export * from "./recording-email-service";
 export { default as renderEmail } from "./src/renderEmail";
